+++ test_records_go/embed.go
@@ -7,6 +7,3 @@ var templatesFS embed.FS
 
 //go:embed static
 var staticFS embed.FS
-
-//go:embed RELEASE_NOTES.md
-var releaseNotesData []byte